requirements.txt
# LLM / Agents
pyautogen>=0.2.35,<0.3
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1

# Vector store
faiss-cpu>=1.7.4

# Embeddings / NLP
tiktoken>=0.5.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Backend API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Frontend
streamlit>=1.30.0

# Data / validation
pydantic>=2.0.0
python-dotenv>=1.0.0

# Observability
langsmith>=0.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
